fix: 깃허브 검색 로직 분리
